pool: Refactor meta data store interface to allow bulk updates
Motivation:

Often we update several attributes together, yet these are individual
updates in the store and cause syncs to disk every time. Such updates
can already be made atomic by synchronizing on the entry, but there
is no way for the underlying storage to batch together the update to
persistent state.

Modification:

Introduced an update method to MetaDataRecord. The argument is a
callback that is given an updatable version of the meta data record.
The callback may make one or more changes to the record.

This change does not imply that other aspects of the meta data record are not
modifiable through the regular interface. The callback is only for attributes
that are typically updated together. It is even assymetric in the sense that
the linkcount is incremented through the callback, but decrement outside.
Although maybe not aesthetically pleasing, it reflects how the interface is
used.

For convenience, the callback has also access to a number of getters.

This patch doesn't actually utilize this to do any bulk updates to
the persistent state. That will be done in subsequent patches.

Result:

No user visible changes.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9402/