pool: Don't use transient error for a broken file
Motivation:

Opening a broken replica is not a transient error, yet we report that the
replica is locked which usually indicates that retrying will eventually allow
the request to succeed.

Modification:

Use FileCorruptedCacheException instead - we use the same error message in
other places when we try to access a broken replica.

Result:

Minimal user visible impact, but useful for resilience as it needs to know
which errors are transient and which are not. Hence the backport request.

Target: trunk
Require-notes: no
Require-book: no
Request: 2.16
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9403/