script: Do not claim success if meta data conversion failed
Motivation:

Although the dcache script runs with `set -e`, this does not seem
to be preserved in subshells.

As a consequence, failure to convert the meta data of a pool is
follow by output claiming the data was converted successfully.

Modification:

Fail the script if the conversion failed.

Result:

Fixed a problem in which failure to convert pool meta data was
misreported.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.16
Request: 2.15
Request: 2.14
Request: 2.13
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9404/