resilience: fix classification of file incomplete and other errors
Motivation:

When a p2p copy fails because it is trying to access a source
file before the pool repository has completed processing it,
the client receives a "Source pool failure (File is incomplete)"
error.   Currently, because of the presence of "Source pool failure"
in the exception message, resilience classifies this as an error
requiring a retry using a different source.   In reality, however,
this is a transient error which can be retried using the same source.

In addition, relying on the error message for this classification
is fragile.

Modification:

Reclassify the error as retriable so that Resilience simply retries
without attempting to find a different source location.

Other classification errors have been rectified, and
reliance on the message eliminated.

A few small changes in the error post-processing
have been made, and several unit tests adjusted.

Result:

Correct behavior by resilience regarding error conditions.

Target: master
Request: 2.16
Require-notes: yes
Require-book: no
Acked-by: Gerd