resilience: simplify code for handling resilience requests on pools
Motivation:

The pools have a special resilience message handler which responds to request
to enforce the sticky flag on a file and to remove an entry from the repository.
It also adds a broken file listener to repository which notifies resilience
when files are so marked.

However, the code for the message handling is unnecessarily complicated and
somewhat difficult to follow.

Modification:

Simplify the handling of the messages.  The changes allow us to eliminate
the task classes as well.

Result:

Code which is clearer but accomplishes the same purpose.

This is not really a bug, so the change is only for master.

Target: master
Acked-by: Gerd
Require-notes: no
Require-book: no