resilience: fix incomplete behavior for pools intentionally marked "excluded"
Motivation:

Under normal operation, the Resilience service should be expected
to handle gracefully situations where a pool with many files, for
one reason or another, goes off line.  Such an incident, even if
the "grace period" value is set to 0, in initiating a large scan,
should not bring the system down.  Obversely, should the pool
come back on line, any such scan should be (and is) immediately canceled,
and the pool rescheduled for a scan to remove unnecessary copies.
Under most circumstances, the Resilience service should not require
any manual intervention or adjustment in this regard.

Nevertheless, admins may still feel the need to have a safety procedure
whereby pools can be excluded temporarily from being handled by
the Resilience service.  This may be used, for instance, if
one decides to rebalance a resilient pool group (resilience should
be disabled there first).  Such a feature was provided in the
old Replica Manager via the "pool set <pool> OFFLINE" command.

The Resilience redesign intended to keep this feature; however,
its current implementation is both unnecessarily redundant and
confusing, as well as being partly erroneous.  So as it stands,
this feature is actually broken, not just awkward to use:

Currently, there are two separate commands to "deactivate" resilience
on a pool, – on the pool itself,

\s poolname pool suppress resilience on/off

and in the Resilience service,

\s Resilience pool exclude/include <expression>

The reason for the first was to take advantage of the fact that one
could save this setting so it survives a pool restart.  But if
a pool is discovered to be unreachable, one also needs a way to
"shut off" resilience for it (if so desired) inside the Resilience
service; hence, the second command.

The two states as a consequence also had to be kept distinct.
The first required modifying the PoolV2Mode object to carry that
information, as well as the pool configuration code.

For obvious reasons, this situation is suboptimal.

But the exclusion/deactivation of a pool in the new system also does
not behave as one would expect (on the basis of the Replica Manager)
in the sense that, while blocking scans of the pool on state changes
or periodic checks, the locations on that pool are still considered
offline and will result as "missing" if a replica check is initiated
on another pool sharing files with it.  This behavior, at least
from the standpoint of the original, is wrong.

Modification:

1.  The pool suppress resilience command, and corresponding INACTIVE
    state handling have all been eliminated.
2.  The pool exclude command now also marks the pool information in
    such as way as to distinguish between 'readable' and 'countable'
    files; handling of resilience requests takes this into account.
3.  To compensate for the loss of the ability to save state on the pool,
    resilience itself now writes out to a file (during the periodic
    update from pool monitor) any pools that are set to EXCLUDED,
    and reads these back in a (re)start.

Note that the PoolV2Mode changes originally introduced to support
INACTIVE have been left in place in order to be backward
compatible with pools on which this patch is not applied.

Unit tests have been adjusted to reflect the new semantics.

Result:

Simpler/unified control over this feature, and behavior as
expected on the basis of past usage.

Target: master
Request: 2.16
Require-notes: yes
Require-book: yes
Acked-by: Gerd