httpd: check for null Subject in Transfer info
Motivation:

The httpd transfer collector classes periodically send
requests to the login broker.  Under certain conditions
it is possible for the reply to contain incomplete
information.  This is particularly true if the reply is sent
before login or file resolution has completed.

Since the TransferObserver and webadmin mapper classes attempt
to convert the Subject into a UserInfo object using the Subjects
methods, a NPE is generated if the Subject is null.

Modification:

Check for the null subject where UserInfo is set and
provide a UserInfo object with an empty constructor in
that case.

Result:

No NullPointerException.

Affects only master (changes introduced by 9258).

Target: master
Acked-by: Gerd