build: add code-coverage reports
Motivation:

We have unit-tests for checking dCache's behaviour.  It would be useful
to know how much of our code-base is covered.

Modification:

Add JaCoCo support (spin-off/next-gen library from EclEmma).  Binary
code-coverage reports are captured from which XML and HTML reports are
generated.  This is added as the profile 'code-coverage' to avoid
the overhead when the developer is only interested in testing dCache.

This patch also removes a work-around that is no longer needed with
JDK 1.8.0_40 or newer.

Result:

Output on code-coverage is available when 'code-coverage' profile is
enabled.

Target: master
Patch: https://rb.dcache.org/r/9430/
Acked-by: Gerd Behrmann
Request: 2.16
Requires-notes: no
Requires-book: no