cells: Allow local delivery of messages through queue routes
Motivation:

The cells framework has two boolean flags to restrict local and remote delivery
of messages.  The current implementation would skip the use of routes if remote
delivery was disabled (with the exception of topic routes). This causes
messages to named queues to fail to be delivered to local consumers if remote
delivery is disabled. This is the case when the message was received from a
core domain.

Modification:

Restructure the routing loop such that only delivery through routes to domain
targets are suppressed. This is the same logic also used for topic routes.

Result:

Fixed an issue with message delivery to named queues affecting consumers
whose cell name is different from the queue name.

Target: trunk
Require-notes: yes
Require-book: no
Request: 2.16
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9438/