doors: enforce restriction in all PnfsManager operations
Motivation:

When the Restriction concept was brought into dCache, the intention was
that it be enforced in the namespace.  Some messages sent by doors and
transfermanager neglected to include the restriction, leading to dCache
ignoring the read-only nature of some requests.

Modification:

Add restriction to all messages sent from doors and transfermanager.
The system-test is updated to include a readonly user; tests are added
to confirm that the anonymous user (if so configured) and user
'readonly' are read-only.

Result:

Restrictions are enforced and (in some cases) checked.

Target: master
Request: 2.16
Request: 2.15
Acked-by: Gerd Behrmann
Acked-by: Tigran Mkrtchyan

Conflicts:
	modules/dcache-srm/src/main/java/diskCacheV111/srm/dcache/Storage.java