restful-api: make exception and error handling resful
Motivation:

1. The current error and exception handling inside the resful-api rely on the default
error handling in jetty. Unfornately, the side-effect of this is that the error message
are in html format.

2. The NotAuthorizedException response include `401` status code with `WWW-Authenticate`
header field. The cause browser to prompt for login and this is not suitable for
dcache-view (web-app).

Modification:

1. Create error handling package for restful-api, this handles both checked and
unchecked exceptions and convert the exception object to desirable JSON format.

2. Handle and support a request with header "X-Requested-With": "XMLHttpRequest". This
remove the WWW-Authenticate header field causing the browser to prompt for login. Even
though this is not a standard way of responding with `401`, the standard behaviour is
still supported whenever the X-Requested-With header is not set to `XMLHttpRequest`
in the request.

Result:

1. proper error handling in JSON format
2. Prompt login can now be averted in dcache-veiw

Target: trunk
Require-notes: no
Require-book: no
Request: 2.16
Acked-by: Gerd Behrmann <behrmann@ndgf.org>

Review at https://rb.dcache.org/r/9457/