alarms: fix NPE in type setter
Motivation:

If the type of the alarm is not set, the server should
handle it by checking to see if it matches a definition
and if not, set it to GENERIC.

This is done during the conversion of the logging event
into a database object.  However, the original event
is what is passed back to the email appender.

If the type on this event is undefined, the email send
should not fail.

Prior to 2.13, it would seem that this did not happen
because the MDC property was being set on even the
typeless events.  This is no longer the case, and
trying to send a generic event via email results in
an NPE.

Modification:

There is really no need to rederive the type from
the event MDC or the alarm again, since the type
has already been determined by the conversion.
Just set that type on the event as well.

Result:

No NPE preventing both the email and the addition of
the alarm to the database.
Request: 2.16
Request: 2.15
Request: 2.14
Request: 2.13
Require-notes: yes
Require-book: no
Fixes: #2578
Acked-by: Gerd