pool: Expose URI rather than File to NearlineStorage implementations
Motivation:

To allow MetaDataStore implementation that do not store files
on the OS file system.

Modification:

Change MetaDataRecord#getDataFile to return a URI rather than a File.

The FlushRequest and StageRequest interfaces are extended with an
additional method, getReplicaUri, to return the URI identifying the
data file.

The existing getFile method is marked deprecated. It continues to
work for pools using the default file system provider and thus we
maintain compatibility with third party nearline storage providers.

The processing of OpenFlags had to be moved from WriteHandleImpl
to MetaDataStore#create to avoid that WriteHandleImpl had to access
the file directly.

Nearline storage implementations shipped with dCache are updated
to use the URI.

The script nearline storage provider maintains the existing command line
argument syntax when used with the default file system provider - for other
providers a full URI is provided in the callout instead. This maintains
backwards compatibility with existing HSM scripts.

Other nearline storage providers shipped with dCache will work with
non-default file system providers, but they will not work with a
MetaDataStore not backed by a FileSystemProvider.

Result:

One can implement a custom MetaDataStore with alternate file backends.
If the backend store is exposed as a FileSystemProvider, all one needs
to do is to implement the dCache FileStore interface.

The NearlineStore SPI was extended with an additional method to obtain
a URI to the replica. Third party implementations will continue to
work as before, but should be updated to use this method.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>

Reviewed at https://rb.dcache.org/r/9460/