httpd: change condition from numeric inequality to non-null check on TransferInfo value
Motivation:

When transfer info from a door is received, some values may not be set
if there is no mover.  These will then be null (including numeric objects).

One condition check in the TransferObserver1 was erroneously treating
one of these fields as an initialized primitive value.

Modification:

Change the condition to a non-null check.

Result:

No NullPointerException if the value is not set.

Target: master
Request: 2.16
Acked-by: Gerd