resilience: allow file consumer to propagate Exception
Motivation:

The file consumer thread is currently catching Exception.
The rationale for this was to allow for "file ctrl start"
to be used to try to recover functioning.

It is, however, preferable to just allow the error to
propagate, since the Runtime exceptions represent bugs
anyway.

Modification:

Remove the catch clause from the run method.

Result:

Clearer indication of bug in log, and fail-fast
behavior.

Target: master
Request: 2.16
Acked-by: Gerd