resilience: remove pool from operation table when it has been removed from a resilient group
Motivation:

The pool operation list in Resilience exposes only potential operations on resilient pools.
If a pool should be removed from a resilient group, it should no longer appear in this list.

Modification:

When a pool scan completes, check to see if the pool is still resilient before
putting it back on the idle queue.

Note that when a pool is removed from a resilient group, the pool is scheduled to
be scanned, so the easiest way to ensure map consistency is lazily, rather than
checking the map periodically.

Result:

Stray pools which were once but are no longer resilient no longer stick around in
the operation map and thus disappear from the "pool ls" listing.

Target: master
Request: 2.16
Acked-by: Gerd