resilience: remove restriction on storage unit linkage
Motivation:

The pool selection unit decorator currently enforces a constraint
on resilient storage units preventing them from being linked to
non-resilient pool groups.

This constraint is unnecessary, however, as resilience only cares
about the semantics of the unit within a given resilient group,
not outside it.

Modification:

Remove the constraint checking.

Result:

There is nothing preventing a resilient storage unit from being
linked to a non-resilient pool group.

This is not an improvement, it is actually incorrect.  It should
be backported.

(I will adjust the documentation on the wiki.)

Target: master
Request: 2.16
Acked-by: Gerd