resilience: fix several related bugs in transition checking when scheduling scans
Motivation:

When a pool changes state, resilience uses an implicit transition table to
determine whether a scan is necessary or not.

However, there are several other paths leading to scanning.

1.  The periodic pool sweeps.
2.  A change in pool groups or storage units.
3.  A manually initiated scan.

Currently, 2) and 3) share the same logic.  This is incorrect.  Except
for pools which are not yet initialized, 3) should never be rejected
(it is a forced scan in the truest sense).  At the same time, there
are distinct differences between whether 2) has to do with a pool
or pool group change or a storage unit change.

Modification:

Distinguish properly between the various cases.  This means that
considerations of whether the pool is down and has already been
scanned, or is excluded, should be bypassed for (3), and also
for a redistribution of pool tag constraints on the storage unit,
but should be enforced otherwise.  (In the latter case we skip
the checks because this is really a form of 'rebalancing' rather
than the creation of missing copies or the removal of extra ones.)

Result:

Correct scan scheduling behavior.

Target: master
Request: 2.16
Acked-by: Gerd