resilience: fix the way removal of a pool from a resilient group is handled
Motivation:

When a pool is removed from a resilient group, the resilience system
should take appropriate action.  This depends upon the (altered)
composition of the pool group and the presence of replicas on other
pools in the group.

If replicas held by the removed group are still present on other
pools in the group, then an attempt to compensate by making the
missing copy elsewhere (just as if the pool were DOWN) should occur.

When, however, all copies of that replica have been removed from
the group (because other pools have been removed), then no action
should occur.

These conditions are both currently being handled, but in a manner
which is fragile.  The latter, in particular, is checking an enum
value which may not be present because the operation is not being
properly updated (another patch will be posted that fixes the more
general issue which affects other areas of the code as well) when
another request to handle the same pnfsid occurs.  It is also
neglecting to do a relevant check earlier (which could
avoid unnecessarily creating or updating an operation to begin with).

Modification:

Eliminate the dependency on the REMOVE value of the SelectionAction
enum (a later patch will eliminate this enum entirely).  Check,
instead, the intersection of attribute locations with the pools
that are actually members of the group in question.  The FileUpdate
now pre-empts the operation when this intersection is null; this
is rechecked in the later verification method (instead of the
suppressAlarm flag).  Some refactoring of the PoolInfoMap methods was
done to make the various calls to find countable, readable and
valid members DRY.

Result:

The remove triggers consistent behavior which survives incoming
operation updates because it depends only on the current state
of the pool group.

Target: master
Request: 2.16
Acked-by: Gerd