resilience: fix error in file operation updating
Motivation:

File operations run concurrently, but are serialized with respect to
pnfsid.  For any logical file, only one copy or remove is thus allowed
to run at a given time.

When several pool scans intersect over their sets of pnfsids, the
first scan to have created the file operation remains the "parent",
but the operation is updated.  Currently, this merely involved
incrementing the operation count, i.e., the number of times it
should attempt to run.

This, however, neglects passing on some important data which
may be held by subsequent scan calls, in particular, whether
the scan involves a specific storage unit, and whether it
also forces the checking of the source sticky bit.  In addition,
in the case of tag distribution, the count calculation done
during the preliminary phase must also be reflected.

Modification:

Change the update call to incorporate the updating of these
attributes/values.

Note that the dependency on MODIFY for determining whether
the operation involves the file's storage unit has been
eliminated in favor of just checking that the unit value
is set.

Result:

Correct behavior on file operation update.

Target: master
Request: 2.16
Acked-by: Gerd