cell: Only implement CellSetupProvider if needed
Motivation:

AbstractCellComponent implements CellSetupProvider to provide empty default
implementations of the three methods. This is to avoid lingering the code with
empty methods for the common case in which only a single or two of those
methods are needed for a particular bean.

With Java 8 there is a better way to achieve this and it allows us to reduce
how many beans implement the CellSetupProvider interface and thus how many
beans are exposed to setup files.

Modification:

Don't implement CellSetupProvider in AbstractCellComponent. Instead default
implementations are provided in the interface. When needed, the interface
is explicitly implemented in beans.

Result:

No user visible changes, except if one misused the cell setup files to
provide non-setup commands.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9468/