pool: Eliminate thread in mover scheduler
Motivation:

A mover scheduler has a single thread to submit movers for execution. There are
however only three cases in which movers are started: When a new mover is submitted
and there is a free slot, when a mover finished and when increasing the number of
free slots.

The first of these cases already bypasses the scheduler thread by letting the
calling thread start the mover directly.

Modification:

Eliminate the thread entierly by letting the postprocessing thread submit the next
mover. Also when increasing the number of job slots, the calling thread fills those
slots if movers are queued.

This has the added benefit of removing a minor race in which the scheduler
thread would acquire a job slot even if no jobs are queued - this resulted in
the inability to immediately "disable" a mover queue by setting the limit to
zero as scheduler thread would alway execute one more mover for the slot it
already acquired.

Result:

Not obvious user visible changes.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>

Reviewed at https://rb.dcache.org/r/9479/