pool: avoid NPE when querying status of a 3rd-party HTTP transfer
Motivation:

An NullPointerException is thrown if the transfer's status is queried
before the transfer is initiated.  This can happen if the pool queues
the request and the transfermanager queries the current status of the
transfer.

Modification:

Update code to be robust against the MoveChannel not existing.

Result:

No more NPE

Target: master
Patch: https://rb.dcache.org/r/9483/
Acked-by: Gerd Behrmann
Fixes: #2596
Request: 2.16
Request: 2.15
Request: 2.14
Request: 2.13