cells: Preserve CommandThrowableException
Motivation:

There is a difference between e.g. CommandSyntaxException and a
CommandThrowableException wrapping a CommandSyntaxException: In the former case
there was a syntax error in the command being issued, while in the latter case
the command was ok, but as part of its execution there was a syntax error (e.g.
a reload command experiencing a syntax error in the setup file being loaded).

Modification:

Avoid unwrapping the CommandThrowableException and output the message text
of the wrapper exception rather than the cause. In all cases in which we
create CommandThrowableExceptions, the message test contains the message
of the cause too.

Result:

Hopefully better error messages.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9486/