cells: Check setup file syntax before applying it to the running cell
Motivation:

When issuing the reload command in a cell, the setup file is applied
directly to the running cell. If the setup file contains a syntax error,
the cell is typically left in an inconsistent state.

Modification:

Create mock objecs of all CellSetupProviders and apply the setup file to those
mock objects before applying it to the actual cell. In case of failure to apply
the setup file to the mock objects, the running cell is not affected.

The CellSetupProvider interface is extended with a mock method. By default this
relies upon the default contstructor of the setup provider to create a mock
object, but in a few cases additional dependencies have to be injected for the
setup provider to be able to parse the setup file.

Result:

The 'reload' command of cells like PoolManager and pools has been refined to
check the syntax of the setup file before reloading it. In case of syntax errors,
the running cell is left untouched.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9487/