SrmShell.java: changed 'rw--' to 'rw-' in case TPermissionMode._RW
I was searching for directory right code to investigate why we have a directory that spontaneously changes from 755 to 770 every now and then. Then I stumbled upon what appears to be a coding typo. I have no idea if it is related to our bug, but I'd like to bring it to your attention anyway.