pool: Cancel permanent migration jobs when reload pool setup file
Motivation:

Permanent migration jobs are persisted in the pool setup file.

Modification:

Cancel existing permanent migration jobs to prevent duplication. Allow
migration jobs that are initializing to be cancelled and silently ignore
a request to cancel if the job is already cancelled or in another final
state. Renumber cancelling jobs with a custom id when a new job with
the same id is being created (allows the setup to be loaded even when the
old job is still cancelling).

Result:

Fixed a problem causing permanent migration jobs to be duplicated
if the pool setup file was reloaded.

Target: trunk
Request: 2.16
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9502/