pool: Adjust failure semantics for empty client uploads
Motivation:

When a client upload fails, this operation is not retriable as the pool marks
the file as final as soon as the write mover terminates.  This is to prevent
that different versions of the file are created on different pools.

Modification:

We may relax the requirement if the client didn't actually write any data to
the pool yet - either because it never connected to the mover or because no
data was written over the data channel. In such cases leaving the file in its
virgin state is no different than if we had never managed to create the mover
in the first place.

This patch thus changes the semantics on failed upload from a client such that
empty replicas are deleted rather than being registered in the name space. The
recovery procedure upon pool restart similarly deletes an empty client upload
if we haven't yet registered it in the name space.

The name space entry is not deleted.

This change will allow us to safely retry pool selection and recreate the mover
on another pool. We cannot make that change until after the next golden release
due to compatibility requirements with old pools.

Result:

Adjusted pool failure semantics such that empty replicas on upload failures are
deleted rather than marked broken.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9506/