frontend/dcache-restful-api: remove www-authenticate header
Motivation:

The inconsistency in browser behaviour on how to handle an ajax call
have some detrimental effect in developing a more reliable and user
friendly web app.

Some browsers set and send all the request headers at once while others
send some basic headers first then later send all the headers set in the
request. Since dcache-rest-api rely on header: X-Requested-with which is
a signal telling the rest-api not to send the Www-authenticate header.
Unfornately some browser don't send this during the first flightof the
request. Therefore a 401 with www-authenticate response header is send
to the browser (which is not desirable).

Modification:

1. Set `www-authenticate` header to an empty string when (only) 401 response
is hit, which in return remove the header.
2. (At least for now) no need for X-Requested-with header.

Result:

This end the problem with prompt login by the browser.

Target: trunk
Request: 2.16
Require-notes: no
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>

Review at https://rb.dcache.org/r/9518/