dcache: Allow UniversalSpringCell to persist setup in zookeeper
Motivation:

UniversalSpringCell allows beans implementing CellSetupProvider to persist the
service setup in a setup file. The setup is persisted using the `save` command
and loaded upon restart and when issuing the `reload` command. If a service
is restarted without first saving the setup, the setup is lost.

In a redundant service, the setup typically needs to be synchronized over
multiple instances.

Modification:

Extends the UniversalSpringCell to allow a setup node to be configured. When
set, the cell synchronizes the service setup with a ZooKeeper node. Whenever
the setup changes, the ZooKeeper node is updated, and whenever the ZooKeeper
node is modified, the service reloads the setup from ZooKeeper.

Enabling the use of a setup node for a service has two consequences:

The first is that the setup survives a service restart even if the setup
isn't saved to the setup file. The setup file will be used to seed the initial
configuration in ZooKeeper, but subsequently one has to explicitly issue the
`reload` command to reload it - a restart will not reload the setup file.

The second is that multiple instances of a service can synchronize their setup
through the ZooKeeper node.

A few words about the synchronization guarantees: Two models were considered:

The preferred approach would have been to execute any setup modifying command
on mock instances of the setup providers, and, if successfull, serialize the
resulting setup to ZooKeeper. The update notification from ZooKeeper would in
turn update the real setup of the service. The benefit would have been that the
local setup would only ever be modified *after* the ZooKeeper node was
modified. This approach wasn't viable as some commands may affect both
persistent and non-persistent aspects and thus could not be applied solely on a
mock of the setup providers (the obvious example is the `migration cancel`
command that can cancel both permanent and non-permanent jobs).

Thus the implementation follows an alternative approach in which the local
setup if modified first and then serialized to ZooKeeper. If the setup command
fails it is assumed that the setup was not modified and the ZooKeeper node is
not updated. The problem is that updating the ZooKeeper node may fail and in
that case the local setup has to be rolled back. In that case the setup command
will be reported as failed, but the change would temporarily have been applied
to the running service. The update in ZooKeeper may fail for various reasons: A
concurrent update to the setup in another instance could mean the local view of
the setup is out of date; or the connection to ZooKeeper could have been lost
and thus the update fails. It was considered to use a distributed lock to
prevent the first case (obtain the lock before applying the change locally),
however in case of a loss of the ZooKeeper connection the node would loose the
lock anyway and we would still need to role back the setup. Given how rare it
is that the setup would be modified concurrently on two different instances, it
was concluded that a distributed lock wasn't worth the trouble.

TL;DR; When issuing a setup modifying command, the command may fail due to
synchronization issues with ZooKeeper. In this case the local setup will have
been modified temporarily and then rolled back to the last known setup from
ZooKeeper.

Result:

No user visible changes as no service makes use of the new functionality yet.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>

Reviewed at https://rb.dcache.org/r/9531/