cells: Extend CellLifeCycleAware with setup change notification
Motivation:

We increasingly rely on disseminating the state of a component to other
components. The primary example of this is pool manager.

Modification:

To simplify such synchronization, this patch adds a notification to the
CellLifecycleAware interface. The setupChanged notification is generated
whenever the current setup changes, either because it was reloaded from
an external source or because a setup command was invoked.

The existing pool selection unit specific change listener support is
removed as pool mananger relies on the generic notification instead.

Result:

No user visible changes.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9537/