dcache-restful-api/pinmanager: modify  current QoS fom disk+tape to tape
Motivation:

 According to CDMI/QosS specification  QoS  of a file object  could be changed from disk+tape (nearline+online) to tape (nearline).
In dCache this could be mapped  to unpin command, which unpins a previously pinned file.
 Nevertheless the same  file could be pinned by different  requestors (user, doors for example SRM door).
Therefore even if file is unpinned by the QOS requestor  it still could be pinned by others and the locality of the file will be NEARLINE+ONLINE.

One of the solutions is to check the number of pins for a file.

Modifications

New `PinMangerQueryQosPinsMessage`  and `QueryQosPinsRequestProcessor`  have been added to pinmanager module,
so that we can query the number of pins for a file.

Result

Users can see the change of QoS from disk+tape to tape.

Target: master
Require-book: no
Require-notes: no
Patch: https://rb.dcache.org/r/9566/
Acted-by: Gerd Behrmann <behrmann@ndgf.org>