cells: Refactor message forwarding logic
Motivation:

Cells support specifying a destination path consisting of more than one address
on a message. Such a message should be forwarded by each intermediate cell.

Currently, the logic to advance the position in the destination path is placed
in each service that can forward messages. Further, services that want to inject
an address on the path have to be careful with how the insert and the advance
are ordered.

Modification:

This patch moves the advance into the cell nucleus. Thus by the time the message
is delivered, the position in the destination path already points to the next
receiver.

This implies that the position in the path can be set to be after the last
element (as would be the case when the message is delivered to the last
receiver).

The advantage here is that the higher level code in the services becomes simpler.

Result:

The deliver event in the cell message event log will have a slightly different
destination path as the current position will have been advanced at the time
of message delivery.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>

Reviewed at https://rb.dcache.org/r/9559/